Move React hooks to new entrypoint
BREAKING CHANGE: Hooks are no longer exported from the main entrypoint
