+++ src/index.ts
@@ -1,6 +1,5 @@
 export { InvalidationPolicyCache, makeCachedVar } from "./cache";
 export { ApolloExtensionsClient } from "./client";
-export { useFragment, useFragmentWhere } from './hooks';
 export { InvalidationPolicyCacheAuditor } from "./audit";
 export {
   DefaultPolicyAction,
